Backed out merge c8b187b855 from stevenschlansker.
A ConfigProvider for the same config class can (and must) exists in
multiple instances, e.g. a provider for Foo that reads its values from
the prefix "a" and a provider for Foo that reads its values from the
prefix "b" must exist side by side:

bind(Foo.class).toProvider(ConfigProvider.of("a", Foo.class)).in(Scopes.SINGLETON);
bind(Foo.class).toProvider(ConfigProvider.of("b", Foo.class)).in(Scopes.SINGLETON);

both the config-magic created instances of Foo and the Providers
themselves are separate instances. Therefore, neither Foo nor the
ConfigProvider<Foo> can be singletons in the guice sense.

In fact, the provider is almost created by Guice itself (it has an
arguments c"tor, that is not marked as @Inject-able) but instance-wise
bound. So the @Singleton annotation is actually confusing because it
can not be enforced by guice.